backend/requirements.txt
fastapi
uvicorn
tensorflow
transformers
pydantic
python-multipart
python-dotenv
google-generativeai
